requirements.txt
langchain-openai>=0.2.0
langgraph>=0.5.0
langchain-core>=0.3.0
requests>=2.31.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-mistralai>=0.2.0
langchain-community>=0.3.0
langchain-cohere>=0.3.0
langchain-fireworks>=0.2.0
langchain-together>=0.2.0
langchain-groq>=0.2.0
